requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
joblib>=1.1.0
flask>=2.0.0
flask-wtf>=1.0.0
python-dotenv>=0.19.0
gunicorn>=20.1.0
pyyaml>=6.0.0
pytest>=6.2.0 